TVG-103 Move show requests into new file
